start of reworking edit
Note to self; maybe you should add a column to the pages-table called is_md, and that being true if the checkbox is checked on creation. That way you can store it as md and then convert it on output based on the is_md. This also allows for editing the markdown without needing to convert it back to md.